[mod] Locater method unified in locater.go
