+++ cmd/locater_test.go
@@ -18,3 +18,51 @@ func TestLocater_Normalize(t *testing.T) {
 		t.Fatalf("got: %v want: %v", actual, expected)
 	}
 }
+
+func TestLocater_highlightString(t *testing.T) {
+	s := "/home/vagrant/Program/hoge3/program_boot.pdf"
+	words := []string{"program", "pdf"}
+	actual := highlightString(s, words)
+	p := "<span style=\"background-color:#FFCC00;\">"
+	q := "</span>"
+	expected := "/home/vagrant/" +
+		p + "Program" + q +
+		"/hoge3/program_boot." +
+		p + "pdf" + q
+	if actual != expected {
+		t.Fatalf("got: %v want: %v", actual, expected)
+	}
+}
+
+func TestLocater_CmdGen(t *testing.T) {
+	l := Locater{
+		SearchWords:  []string{"the", "path", "for", "search"},
+		ExcludeWords: []string{"exclude", "paths"},
+		Dbpath:       "../test/mlocatetest.db:../test/mlocatetest1.db",
+	}
+	actual := l.CmdGen()
+	expected := [][]string{
+		{
+			"gocate",
+			"--database",
+			"../test/mlocatetest.db:../test/mlocatetest1.db",
+			"--",
+			"--ignore-case",
+			"--quiet",
+			"--existing",
+			"--nofollow",
+			"--regex",
+			"the.*path.*for.*search",
+		},
+		{"grep", "-ivE", "exclude"},
+		{"grep", "-ivE", "paths"},
+	}
+	t.Logf("expected command: %v, actual command: %v", expected, actual) // Print command
+	for i, e1 := range expected {
+		for j, e2 := range e1 {
+			if actual[i][j] != e2 {
+				t.Fatalf("got: %v want: %v\ncommand: %s", actual[i][j], e2, actual)
+			}
+		}
+	}
+}
